+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.9.1 (Next)
 ============
+* [#820](https://github.com/intridea/grape/pull/820): Added missing method `all_or_none_of` for the DSL parameters - [@natecj](https://github.com/natecj).
 * [#803](https://github.com/intridea/grape/pull/803): Added `all_or_none_of` parameter validator - [@loveltyoic](https://github.com/loveltyoic).
 * [#774](https://github.com/intridea/grape/pull/774): Extended `mutually_exclusive`, `exactly_one_of`, `at_least_one_of` to work inside any kind of group: `requires` or `optional`, `Hash` or `Array` - [@ShPakvel](https://github.com/ShPakvel).
 * [#743](https://github.com/intridea/grape/pull/743): Added `allow_blank` parameter validator to validate non-empty strings - [@elado](https://github.com/elado).